+++ src/main/java/com/vtube/dal/CommentsRepository.java
@@ -7,14 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.vtube.model.Comment;
-import com.vtube.model.Video;
 
 @Repository
 public interface CommentsRepository extends JpaRepository<Comment, Long>{
 
 	Optional<Comment> findById(Integer commentId);
 
-	List<Comment> findAllBySuperCommentId(Integer commentId);
+	List<Comment> findAllBySuperCommentId(Long commentId);
